refactor(request): extract login utility helpers
